tado: remove weather
Only sensors will be exposed until a decission regarding the lack of humidity
is made.
