+++ homeassistant/components/tado/manifest.json
@@ -7,5 +7,6 @@
   "config_flow": true,
   "homekit": {
     "models": ["tado", "AC02"]
-  }
+  },
+  "dependencies": ["weather"]
 }
